refactor: delete use of pkg directory
