+++ provider/json_rpc_provider_test.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/jarcoal/httpmock"
-	"github.com/pokt-foundation/pocket-go/pkg/mock-client"
+	"github.com/pokt-foundation/pocket-go/mock-client"
 	"github.com/stretchr/testify/require"
 )
