+++ provider/rpc_outputs.go
