+++ provider/samples/query_app.json
