+++ provider/samples/query_apps.json
