+++ provider/samples/query_block.json
