+++ signer/interfaces.go
