+++ signer/key_manager.go
@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 	"errors"
 
-	"github.com/pokt-foundation/pocket-go/pkg/utils"
+	"github.com/pokt-foundation/pocket-go/utils"
 )
 
 var (
