+++ signer/key_manager_test.go
