+++ signer/wallet.go
